+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added support for Pybricks Protocol v1.1.0.
 
+### Fixed
+- Fixed `pybricksdev ble run` not working with BOOST Move hub.
+
 ## [1.0.0-alpha.10] - 2021-06-27
 ### Added
 - Support for Python 3.9.